+++ lib_comfyui/comfyui_adapter.py
@@ -4,9 +4,11 @@ import importlib
 from modules import shared
 from torch import multiprocessing
 from modules import script_callbacks
-from lib_comfyui import async_comfyui_loader, webui_settings, argv_conversion
+from lib_comfyui import async_comfyui_loader, webui_settings, comfyui_reverse_proxy
 importlib.reload(webui_settings)
 importlib.reload(async_comfyui_loader)
+importlib.reload(comfyui_reverse_proxy)
+should_share = '--share' in sys.argv or '--ngrok' in sys.argv
 
 
 process = None
@@ -17,9 +19,8 @@ def start():
     if not os.path.exists(install_location):
         return
 
-    should_share = '--share' in sys.argv or '--ngrok' in sys.argv
     if should_share:
-        start_comfyui_localtunnel()
+        comfyui_reverse_proxy.start()
 
     model_queue = multiprocessing.Queue()
     start_comfyui_process(model_queue, install_location)
@@ -46,45 +47,13 @@ def start_comfyui_process(model_queue, install_location):
         sys.path.extend(original_sys_path)
 
 
-def start_comfyui_localtunnel():
-    comfyui_argv = list(sys.argv)
-    argv_conversion.set_comfyui_argv(comfyui_argv)
-    port = comfyui_argv[comfyui_argv.index('--port') + 1]
-
-    import shutil
-    local_tunnel = shutil.which('lt')
-    if local_tunnel is None:
-        return
-
-    import subprocess
-    import threading
-    import time
-    import socket
-
-    def on_create_local_tunnel(int_port):
-        while True:
-            time.sleep(0.5)
-            sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-            result = sock.connect_ex(('127.0.0.1', int_port))
-            if result == 0:
-                break
-            sock.close()
-
-        print("Launching localtunnel...")
-        p = subprocess.Popen([local_tunnel, "--port", port, "--local-host", '127.0.0.1'], stdout=subprocess.PIPE)
-        for line in p.stdout:
-            line = line.decode()
-            if line.startswith('your url is:'):
-                webui_settings.set_comfyui_url(line.split('your url is:')[1].strip())
-            print(line, end='')
-
-    threading.Thread(target=on_create_local_tunnel, args=(int(port),), daemon=True).start()
-
-
 def stop():
     global process
     if process is None:
         return
+
+    if should_share:
+        comfyui_reverse_proxy.stop()
 
     process.terminate()
     process = None
